+++ CHANGELOG.md
@@ -20,6 +20,8 @@
       - It will also try to auto resolve conflicts by LR(1) peeking.
     - Add `ParserBuilder.checkAll` to do all necessary checks.
     - Replace `dataReducer` with `LR.reducer`.
+- Manager
+  - Manager will use `ILexer` instead of `Lexer` in constructor.
 - Optimize document / comments.
 - Optimize code / performance.
 - More test code.